Move Bad Words scanning script logic to dmscripts module
- Keeps os.path checks and arg parsing in top level script
- Refactors some of the functions to be more concise
- Pass logger through to functions where needed
